refactor : directory structure modify
