+++ main.scss
@@ -1,3 +1,3 @@
-@import './scss/common.scss';
-@import './scss/header.scss';
-@import './scss/index.scss';
+@import './src/scss/common.scss';
+@import './src/scss/header.scss';
+@import './src/scss/index.scss';
